backend/requirements.txt
flask
flask-cors
firebase-admin
google-genai
python-dotenv
reportlab
gunicorn
numpy
